+++ backend/src/main/java/com/gridsandcircles/gc_coffee/member/controller/MemberController.java
@@ -4,8 +4,8 @@ import com.gridsandcircles.gc_coffee.global.dto.ApiResponse;
 import com.gridsandcircles.gc_coffee.member.dto.MemberRequest;
 import com.gridsandcircles.gc_coffee.member.dto.MemberResponse;
 import com.gridsandcircles.gc_coffee.member.service.MemberService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,11 +19,9 @@ public class MemberController {
     private final MemberService memberservice;
 
     @PostMapping
-    public ResponseEntity<ApiResponse<MemberResponse>> createOrFindMember(
-            @RequestBody MemberRequest request){
+    public ApiResponse<MemberResponse> createOrFindMember(
+            @RequestBody @Valid MemberRequest request){
         MemberResponse response = memberservice.findOrCreateMember(request);
-
-        return ResponseEntity.ok(ApiResponse.ok(response));
+        return ApiResponse.ok(response);
     }
-
 }
