+++ backend/src/main/java/com/gridsandcircles/gc_coffee/member/dto/MemberRequest.java
@@ -1,6 +1,11 @@
 package com.gridsandcircles.gc_coffee.member.dto;
 
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
 public record MemberRequest(
+        @NotBlank(message = "이메일은 필수입니다")
+        @Email(message = "올바른 이메일 형식이 아닙니다")
         String email
 ) {
 }
