+++ backend/src/main/java/com/gridsandcircles/gc_coffee/order/controller/OrderController.java
@@ -1,16 +1,12 @@
 package com.gridsandcircles.gc_coffee.order.controller;
 
-import com.gridsandcircles.gc_coffee.entity.Order;
 import com.gridsandcircles.gc_coffee.global.dto.ApiResponse;
 import com.gridsandcircles.gc_coffee.order.dto.OrderCreateRequest;
 import com.gridsandcircles.gc_coffee.order.dto.OrderDetailResponse;
-import com.gridsandcircles.gc_coffee.order.dto.OrderListResponse;
 import com.gridsandcircles.gc_coffee.order.service.OrderService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/orders")
@@ -30,16 +26,5 @@ public class OrderController {
     public ApiResponse<OrderDetailResponse> getOrderDetails(@PathVariable("orderId") Long orderId) {
         OrderDetailResponse response = orderService.getOrderDetails(orderId);
         return ApiResponse.ok(response);
-    }
-
-    @GetMapping
-    public ApiResponse<List<OrderListResponse>> list(){
-        List<Order> orderList = orderService.list();
-
-        List<OrderListResponse> orderListResponseList = orderList.reversed().stream()
-                .map(OrderListResponse::new)
-                .toList();
-
-        return ApiResponse.ok(orderListResponseList);
     }
 }
