+++ backend/src/main/java/com/gridsandcircles/gc_coffee/stock/dto/StockResponse.java
@@ -1,6 +1,6 @@
 package com.gridsandcircles.gc_coffee.stock.dto;
 
-public record StockResponseDto(
+public record StockResponse(
         Long productId, //상품 아이디
         int availableStock //현재 남은 재고
 ) {
